requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
imageio~=2.16.0
opencv-python~=4.5.5.62
numpy~=1.22.2
pandas~=1.4.1
tqdm~=4.63.0
argparse~=1.4.0
matplotlib~=3.5.1
scikit-learn~=1.0.2
Pillow~=9.0.1
wandb~=0.12.11
utils~=1.0.1
seaborn~=0.11.2
requests~=2.27.1
thop~=0.0.31.post2005241907
plotly~=5.6.0
torch==1.11.0+cu113
torchvision==0.12.0+cu113

